Split EditorTextViewTests into focused test files
Break the 1348-line monolith into 5 files by concern:
- EditorTextViewTests.swift — init, editing, block splitting (206 lines)
- EditorUndoTests.swift — undo, redo, coalescing (212 lines)
- EditorRenderingTests.swift — markdown rendering, coordinate mapping, recompose (405 lines)
- EditorDocumentTests.swift — document loading (86 lines)
- EditorIndentationTests.swift — tab/shift-tab indent unit + integration (400 lines)
